backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
python-dotenv
PyJWT
httpx
firebase-admin
selenium
beautifulsoup4
requests
